+++ __tests__/core.js
@@ -294,7 +294,7 @@ test(
             } = data;
 
             t.false(errored, 'no errored');
-            t.is(results[0].deprecations.length, 3, 'flags no deprecations');
+            t.is(results[0].deprecations.length, 2, 'flags no deprecations');
             t.is(results[0].invalidOptionWarnings.length, 0, 'flags no invalid option warnings');
             t.is(results[0].warnings.length, 0, 'flags no warnings');